% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmaps-GMT.R
\name{addcoast}
\alias{addcoast}
\title{Adding coast contour to existing map}
\usage{
addcoast(reso = 3, coastcolor = rgb(0.37, 0.35, 0.3), ...)
}
\arguments{
\item{reso}{is the coastline resolution, from 1 (full resolution) to  5 (crude). Recommended value for basin wide maps is 4.}
}
\description{
Adding coast contour to existing map
}
\examples{
#TBD 
}
