% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmaps-GMT.R
\name{nice.map.over}
\alias{nice.map.over}
\title{Plotting map with nice labels and formatting on the existing plot}
\usage{
nice.map.over(X, Y, reso = 5, grid = TRUE, col = light.grey, cex = 1)
}
\arguments{
\item{X, Y}{are the vectors of longitude and latitude, can be of length two, with map borders.}

\item{reso}{is the map resolution, from 1 (full resolution) to  5 (crude). Recommended value for basin wide maps is 4.}

\item{grid}{is boolean. If TRUE, then the grid will be plotted at intervals determined by function 'axis'.}

\item{col}{is the color of the land.}

\item{cex}{defines the size of the axis labels passed as cex.axis in function 'par'}

\item{col.sea}{is the color of the sea.}

\item{mar}{are the user margins passed to function 'par'}
}
\description{
Plotting map with nice labels and formatting on the existing plot
}
\examples{
#Pacific-wide map with default settings 
nice.map(120:290,-50:50)
nice.map(c(120,290),c(-50,50)) # same output
}
