% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmaps-GMT.R
\name{nice.map}
\alias{nice.map}
\title{Plotting map with nice labels and formatting}
\usage{
nice.map(
  X,
  Y,
  reso = 4,
  grid = TRUE,
  col = "lightgrey",
  col.sea = "white",
  mar = c(4, 6, 6, 2),
  cex = 1
)
}
\arguments{
\item{X, Y}{are the vectors of longitude and latitude, can be of length two, with map borders.}

\item{reso}{is the map resolution, from 1 (full resolution) to  5 (crude). Recommended value for basin wide maps is 4.}

\item{grid}{is boolean. If TRUE, then the grid will be plotted at intervals determined by function 'axis'.}

\item{col}{is the color of the land.}

\item{col.sea}{is the color of the sea.}

\item{mar}{are the user margins passed to function 'par'}

\item{cex}{defines the size of the axis labels passed as cex.axis in function 'par'}
}
\description{
Plotting map with nice labels and formatting
}
\examples{
#Pacific-wide map with default settings 
nice.map(120:290,-50:50)
nice.map(c(120,290),c(-50,50)) # same output
}
